% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_core_modified.R
\name{alpha_core_modified}
\alias{alpha_core_modified}
\title{Extract Cluster Cores From Fuzzy Clustering}
\usage{
alpha_core_modified(eset, clusters, min.acore = 0.5, reference, names_col)
}
\arguments{
\item{eset}{An expression set used for fuzzy clustering}

\item{clusters}{A set of clusters from fuzzy clustering, with protein names
and membership values}

\item{min.acore}{Cluster membership threshold, i.e. the minimum membership
value for proteins to be included in the alpha core for each cluster}

\item{reference}{Reference table for converting protein names}

\item{names_col}{Column in reference table to use for merging}
}
\value{
A data frame containing the proteins assigned to the alpha core
for each cluster
}
\description{
Extract Cluster Cores From Fuzzy Clustering
}
