% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_missingness.R
\name{assign_missingness}
\alias{assign_missingness}
\title{Assign Missingness to Proteomics Samples}
\usage{
assign_missingness(
  data,
  names_col,
  sam_col,
  val_col,
  rep_col,
  ref,
  MAR_thres = 2,
  MNAR_thres = 0
)
}
\arguments{
\item{data}{A data frame in long format}

\item{names_col}{Column containing protein names}

\item{sam_col}{Column containing samples}

\item{val_col}{Column containing intensity values}

\item{rep_col}{Column containing replicates}

\item{ref}{Reference sample}

\item{MAR_thres}{Minimum samples for MAR (default 2)}

\item{MNAR_thres}{Maximum samples for MNAR (default 0)}
}
\value{
A data frame in long format, with missingness assigned for
each comparison
}
\description{
Assign Missingness to Proteomics Samples
}
