% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_enrichGO_enhanced.R
\name{clusters_enrichGO_enhanced}
\alias{clusters_enrichGO_enhanced}
\title{Perform Gene Ontology Analysis On Clusters}
\usage{
clusters_enrichGO_enhanced(
  data,
  OrgDb,
  keyType = "ORF",
  names_col = keyType,
  clus_col,
  ont = "all",
  reference,
  simplify = TRUE,
  evaluate = TRUE,
  convert = TRUE,
  toType = "GENENAME"
)
}
\arguments{
\item{data}{Data frame containing protein names and cluster assignments}

\item{OrgDb}{Organism database to use}

\item{keyType}{Format of the names (default ORF), as defined by
clusterProfiler}

\item{names_col}{Column containing the names (defaults to keyType)}

\item{clus_col}{Column containing the cluster assignments}

\item{ont}{The ontologies to analyse (BP, CC and/or MF; defaults to "all")}

\item{reference}{Data frame containing the background list in multiple
formats}

\item{simplify}{Whether to simplify the GO result}

\item{evaluate}{Whether to evaluate the GO result}

\item{convert}{Whether to convert the names in the GO result}

\item{toType}{Format to convert the names to (if convert = TRUE)}
}
\value{
A data frame containing enriched GO terms for each cluster
}
\description{
Perform Gene Ontology Analysis On Clusters
}
