% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters_enrichGO_plot.R
\name{clusters_enrichGO_plot}
\alias{clusters_enrichGO_plot}
\title{Plot Gene Ontology Analysis Result From Clustering}
\usage{
clusters_enrichGO_plot(go_result, save = FALSE, filename = NULL, width = 160)
}
\arguments{
\item{go_result}{Data frame containing enriched GO terms}

\item{save}{Whether to save the plot (default = FALSE)}

\item{filename}{The filename to save to}

\item{width}{The width of the saved plot (in mm)}
}
\value{
A plot of enriched GO terms in clusters, coloured by adjusted
p-value
}
\description{
Plot Gene Ontology Analysis Result From Clustering
}
