% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaR.R
\name{deltaR}
\alias{deltaR}
\title{Calculate Log2 Differences Between Samples}
\usage{
deltaR(data, val_col, key_col, delta = NULL)
}
\arguments{
\item{data}{A data frame in long format. The conditions you wish to
compare must be present in a separate column to the other sample
information.}

\item{val_col}{Column containing intensity values}

\item{key_col}{Column containing the conditions to compare}

\item{delta}{Vector specifying the conditions to compare (default is all)}
}
\value{
A data frame containing log2 differences for the comparisons of
interest
}
\description{
Calculate Log2 Differences Between Samples
}
