% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGO_plot.R
\name{enrichGO_plot}
\alias{enrichGO_plot}
\title{Plot Gene Ontology Analysis Result}
\usage{
enrichGO_plot(go_result, save = FALSE, filename = NULL, width = 160)
}
\arguments{
\item{go_result}{Data frame containing output from enrichGO_enhanced}

\item{save}{Whether to save the plot (default = FALSE)}

\item{filename}{The filename to save to}

\item{width}{The width of the saved plot (in mm)}
}
\value{
A plot of gene ratios for enriched GO terms, coloured by
adjusted p-value
}
\description{
Plot Gene Ontology Analysis Result
}
