% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfuzz_prep.R
\name{mfuzz_prep}
\alias{mfuzz_prep}
\title{Prepare Data for Fuzzy Clustering}
\usage{
mfuzz_prep(data, thres = 0, replace = FALSE, mode = "mean", ...)
}
\arguments{
\item{data}{A data frame in wide format. Columns should be samples, and rows
should be proteins.}

\item{thres}{The proportion of missing values that is allowed (default 0)}

\item{replace}{Whether NAs should be replaced}

\item{mode}{The method for replacing NAs (default "mean")}

\item{...}{Other parameters}
}
\value{
A standardised expression set ready for fuzzy clustering using Mfuzz
}
\description{
Prepare Data for Fuzzy Clustering
}
