% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_enhanced.R
\name{pca_enhanced}
\alias{pca_enhanced}
\title{Produce PCA Plots to Compare Proteomics Samples}
\usage{
pca_enhanced(
  data,
  samples = NULL,
  names_col,
  sam_col,
  val_col,
  format = NULL,
  colour_col = NULL,
  shape_col = NULL,
  label_col = NULL
)
}
\arguments{
\item{data}{A data frame in long format}

\item{samples}{List of samples to count (defaults to all)}

\item{names_col}{Column containing protein names}

\item{sam_col}{Column containing samples}

\item{val_col}{Column containing intensity values}

\item{format}{Format of the sample names, as a vector}

\item{colour_col}{Column to use for colouring points (defaults to sample,
otherwise must be within format)}

\item{shape_col}{Column to use for determining shape of points (defaults
to NA, otherwise must be within format)}

\item{label_col}{Column to use for determining label of points (defaults
to sample, otherwise must be within format)}
}
\value{
A plot showing the first two PCs
}
\description{
Produce PCA Plots to Compare Proteomics Samples
}
