% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fuzzy_clusters.R
\name{plot_fuzzy_clusters}
\alias{plot_fuzzy_clusters}
\title{Plot Output from Fuzzy Clustering}
\usage{
plot_fuzzy_clusters(
  eset,
  clusters,
  samples,
  plot_background = TRUE,
  plot_centres = TRUE,
  mem_thres = 0.5,
  lwd = 0.75,
  major_x = NULL,
  minor_x = NULL,
  nrow = NULL,
  ncol = NULL,
  ...
)
}
\arguments{
\item{eset}{A standardised expression set}

\item{clusters}{A list containing the output from fuzzy clustering}

\item{samples}{A list of the samples analysed}

\item{plot_background}{Whether or not to plot the background data for each
cluster (default TRUE)}

\item{plot_centres}{Whether or not to plot the cluster centres (default TRUE)}

\item{mem_thres}{Cluster membership threshold, i.e. the minimum membership
value for proteins to be plotted in each cluster}

\item{lwd}{Line thickness (default 0.75)}

\item{major_x}{Major divisions to plot along the x axis. Samples are plotted
on whole numbers so divisions should be at 0.5 increments.}

\item{minor_x}{Minor divisions to plot along the x axis. Samples are plotted
on whole numbers so divisions should be at 0.5 increments.}

\item{nrow}{Number of rows to plot in facet_wrap}

\item{ncol}{Number of columns to plot in facet_wrap}

\item{...}{}
}
\value{
Line plots showing the proteins assigned to each cluster along with
the background and cluster centres
}
\description{
Plot Output from Fuzzy Clustering
}
