% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_count.R
\name{protein_count}
\alias{protein_count}
\title{Determine & Plot Number of Proteins Per Sample}
\usage{
protein_count(
  data,
  samples = NULL,
  names_col,
  sam_col,
  val_col,
  plot = FALSE,
  format = NULL,
  fill = NULL
)
}
\arguments{
\item{data}{A data frame in long format}

\item{samples}{List of samples to count (defaults to all)}

\item{names_col}{Column containing protein names}

\item{sam_col}{Column containing samples}

\item{val_col}{Column containing intensity values}

\item{plot}{Whether to the output should be a plot or not (default FALSE)}

\item{format}{Format of the sample names, as a vector}

\item{fill}{Column to use for colouring plot (must be within format)}
}
\value{
A plot and/or data frame showing the number of proteins detected in each sample
}
\description{
Determine & Plot Number of Proteins Per Sample
}
