% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_correlations.R
\name{replicate_correlations}
\alias{replicate_correlations}
\title{Calculate Pairwise Correlations Between Replicates}
\usage{
replicate_correlations(
  data,
  samples = NULL,
  protein_id,
  sample,
  value,
  replicate,
  nrow = NULL,
  ncol = NULL,
  limits = NULL
)
}
\arguments{
\item{data}{A data frame in long format}

\item{samples}{List of samples to assess (defaults to all)}

\item{protein_id}{Column containing protein names}

\item{sample}{Column containing samples}

\item{value}{Column containing intensity values}

\item{replicate}{Column containing replicates}

\item{nrow}{Number of rows for facet_wrap}

\item{ncol}{Number of columns for facet_wrap}

\item{limits}{Limits for colour scale}
}
\value{
A plot showing pairwise linear correlations between replicates
of the same sample
}
\description{
Calculate Pairwise Correlations Between Replicates
}
