% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribosome_engagement.R
\name{ribosome_engagement}
\alias{ribosome_engagement}
\title{Sum Up Fractions to Calculate Ribosome Engagement}
\usage{
ribosome_engagement(
  data,
  val_col,
  key_col,
  ribosome_fractions,
  total = NULL,
  normalise = TRUE
)
}
\arguments{
\item{data}{A data frame in long format. The "fractions" must be present
in a separate column to the other sample information.}

\item{val_col}{Column containing intensity values}

\item{key_col}{Column containing fractions}

\item{ribosome_fractions}{The ribosome fractions to sum up}

\item{total}{The "total" fraction to use for normalisation}

\item{normalise}{Whether or not to normalise the ribosome_engagement score
to the total (default is TRUE)}
}
\value{
A data frame containing (normalised) ribosome engagement scores
and totals
}
\description{
Sum Up Fractions to Calculate Ribosome Engagement
}
