% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaR_zscore.R
\name{zscore}
\alias{zscore}
\title{Calculate Z-Scores & P-Values Based on a Reference Population}
\usage{
zscore(
  data,
  samples = NULL,
  comparisons = NULL,
  sam_col,
  val_col,
  key_col,
  ref_col,
  reference,
  alpha = 0.05
)
}
\arguments{
\item{data}{A data frame in long format}

\item{samples}{Vector containing the samples to test (if only a subset of
the possible values)}

\item{comparisons}{Vector containing the comparisons to test (if only a
subset of the possible values)}

\item{sam_col}{Column containing samples}

\item{val_col}{Column containing delta values}

\item{key_col}{Column containing the comparisons to test}

\item{ref_col}{Column containing the protein type (used to determine
the reference population)}

\item{reference}{Indicator of which proteins to use as the reference}

\item{alpha}{Significance level (default 0.05)}
}
\value{
A data frame containing z-score and p-value columns in addition
to the input columns
}
\description{
Calculate Z-Scores & P-Values Based on a Reference Population
}
