% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_completeness.R
\name{data_completeness}
\alias{data_completeness}
\title{Evaluate Completeness of Proteomics Data}
\usage{
data_completeness(
  data,
  samples = NULL,
  names_col,
  sam_rep_col,
  sam_col,
  comp_col,
  imp_col,
  ref,
  ref_comp,
  thresholds = c(0.9, 0.8, 0.6)
)
}
\arguments{
\item{data}{A data frame in long format}

\item{samples}{List of samples to assess (defaults to all)}

\item{names_col}{Column containing protein names}

\item{sam_rep_col}{Column containing samples (including replicates)}

\item{sam_col}{Column containing samples (excluding replicates)}

\item{comp_col}{Column containing comparisons}

\item{imp_col}{Column containing whether value is imputed or not}

\item{ref}{Reference sample (excluding replicate)}

\item{ref_comp}{Reference comparison}

\item{thresholds}{Thresholds defining quality categories (must be three
values; default 0.9, 0.8, 0.6)}
}
\value{
A data frame, indicating the proportion of sample values that
are imputed for each protein
}
\description{
Evaluate Completeness of Proteomics Data
}
