% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_conversion.R
\name{names_conversion}
\alias{names_conversion}
\title{Gene & Protein Name Conversion}
\usage{
names_conversion(
  data,
  names_col,
  fromType = "UNIPROT",
  toType = c("GENENAME", "ORF"),
  OrgDb
)
}
\arguments{
\item{data}{A data frame}

\item{names_col}{The column containing the names you want to convert}

\item{fromType}{The starting format of the names (default UNIPROT), as defined by clusterProfiler}

\item{toType}{The format(s) to convert the names to (default GENENAME and ORF), as defined by clusterProfiler}

\item{OrgDb}{The organism database to use}
}
\value{
A data frame containing both the original names and the conversions
}
\description{
Gene & Protein Name Conversion
}
