% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_correlations.R
\name{replicate_correlations}
\alias{replicate_correlations}
\title{Calculate Pairwise Correlations Between Replicates}
\usage{
replicate_correlations(
  data,
  samples = NULL,
  sam_col,
  val_col,
  rep_col,
  nrow = NULL,
  ncol = NULL
)
}
\arguments{
\item{data}{A data frame in long format}

\item{samples}{List of samples to assess (defaults to all)}

\item{sam_col}{Column containing samples}

\item{val_col}{Column containing intensity values}

\item{rep_col}{Column containing replicates}

\item{nrow}{Number of rows for facet_wrap}

\item{ncol}{Number of columns for facet_wrap}
}
\value{
A plot showing pairwise linear correlations between replicates
of the same sample
}
\description{
Calculate Pairwise Correlations Between Replicates
}
